refactor: fix the documentation for `For(int)` (#291)
This PR corrects documentation for the `For(int)` method across the Mockolate codebase. The previous documentation incorrectly stated that `For(int)` "limits" or "only executes" a callback/return/throw for a given number of times, when it actually "repeats" the behavior for that number of times. The PR also updates the implementation to directly pass the `times` parameter instead of using a lambda predicate, and fixes several formatting inconsistencies in XML documentation comments.

### Key changes:
- Updated documentation from "Limits...to only execute" to "Repeats" for all `For(int)` method overloads
- Changed implementation calls from `For(x => x < times)` to `For(times)` 
- Refactored the `Callback` class to use `int?` field instead of `Func<int, bool>?` predicate
- Fixed line wrapping and formatting in various XML documentation comments